move components to another lib
