+++ CMakeLists.txt
@@ -6,3 +6,4 @@ add_subdirectory (./services/dummy_http_server)
 add_subdirectory (./services/api_config)
 add_subdirectory (./libraries/http)
 add_subdirectory (./libraries/common)
+add_subdirectory (./libraries/components)
