+++ libraries/components/CMakeLists.txt
@@ -1,10 +1,9 @@
 cmake_minimum_required(VERSION 3.0)
 
-project(lib_common)
+project(lib_components)
 set (CMAKE_CXX_STANDARD 17)
 
-find_package(Threads REQUIRED)
-find_package(Boost COMPONENTS regex REQUIRED)
+find_package(Boost REQUIRED)
 
 # libraries headers
 set(LIBS_DIR ../../libraries)
@@ -12,16 +11,9 @@ set(THIRDPARTY_LIBS ${LIBS_DIR}/thirdparty)
 
 # sources
 set(SOURCES
-    src/config/config.cpp
-    src/config/logging_config.cpp
-    src/logging/logger.cpp
-    src/logging/logging_impl.cpp
-    src/logging/sink_stdout.cpp
-    src/logging/sink_string.cpp
-    src/logging/sink_fs.cpp
-    src/format/format.cpp
-    src/threading/thread_pool.cpp
-    src/utils/errors.cpp
+    src/components_base.cpp
+    src/components_controller.cpp
+    src/components_engine.cpp
 )
 
 # headers
@@ -45,21 +37,25 @@ set(COMPILE_OPTIONS
 add_library(${PROJECT_NAME} STATIC ${SOURCES})
 target_include_directories(${PROJECT_NAME} PRIVATE ${HEADERS})
 target_compile_options(${PROJECT_NAME} PRIVATE ${COMPILE_OPTIONS})
-target_link_libraries(${PROJECT_NAME} Threads::Threads)
+
+# link libs
+target_link_libraries(${PROJECT_NAME} lib_http)
+target_link_libraries(${PROJECT_NAME} lib_common)
 
 # test sources
 set(TEST_SOURCES
-    tests/config.cpp
-    tests/format.cpp
-    tests/log.cpp
+    tests/components.cpp
     tests/main.cpp
-    tests/strong_typedef.cpp
-    tests/utils.cpp
     ${SOURCES}
 )
 
+# build tests
 set(TESTS_NAME ${PROJECT_NAME}_tests)
 add_executable(${TESTS_NAME} ${TEST_SOURCES})
 target_include_directories(${TESTS_NAME} PRIVATE ${HEADERS})
 target_compile_options(${TESTS_NAME} PRIVATE ${COMPILE_OPTIONS})
-target_link_libraries(${TESTS_NAME} Boost::regex Threads::Threads)
+
+# link tests libs
+target_link_libraries(${TESTS_NAME} lib_http)
+target_link_libraries(${TESTS_NAME} lib_common)
+
