+++ libraries/components/src/components_base.cpp
@@ -1,6 +1,6 @@
-#include <common/include/component_base.hpp>
+#include <components/include/component_base.hpp>
 
-namespace common::components {
+namespace components {
 
 ComponentBase::ComponentBase() : is_ready_{false} {}
 ComponentBase::~ComponentBase() {}
@@ -9,4 +9,4 @@ bool ComponentBase::IsReady() const {
     return is_ready_;
 }
 
-} // namespace common::components
+} // namespace components
